pyproject.toml
[project]
name = "saferoute-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet>=2025.12.0",
    "folium>=0.20.0",
    "geopandas>=1.1.2",
    "load-dotenv>=0.1.0",
    "networkx>=3.6.1",
    "openai>=2.21.0",
    "optuna>=4.7.0",
    "osmnx>=2.1.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "shapely>=2.1.2",
    "streamlit>=1.54.0",
    "streamlit-folium>=0.26.1",
]

[tool.uv.workspace]
members = [
    "saferoute-ia",
]
